requirements.txt
boto3==1.4.4
botocore==1.5.4
cloudinary==1.5.0
dj-database-url==0.4.2
Django==1.10.5
django-cors-headers==2.0.0
django-filter==1.0.1
django-s3direct==0.4.8
django-storages==1.5.2
django-storages-redux==1.3.2
django-webpack-loader==0.4.1
djangorestframework==3.5.3
docutils==0.13.1
gunicorn==19.6.0
jmespath==0.9.0
Markdown==2.6.7
psycopg2==2.6.2
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
urllib3==1.20
whitenoise==3.2.3
